+++ src/app/features/cart/models/cart-item.ts
@@ -1,6 +1,12 @@
+import { ArgumentOutOfRangeError } from 'rxjs';
 import { ProductModel } from '../../products/models/product';
 
-export interface CartItem {
-  item: ProductModel;
-  amount: number;
+export class CartItem {
+  constructor(
+    public item: ProductModel,
+    public amount: number) { }
+
+  getTotalItemPrice(): number {
+    return this.amount * this.item.price;
+  };
 }